requirements.txt
requests
beautifulsoup4
python-dotenv
playwright
streamlit
plotly
pandas
sqlite3